gc: Reverted to only consider ckmeta/ckdata done if not mutated
Two main reasons:

1. If we mount without ckprogs, we do actually have a pretty decent hole
   here where data can be written with errors and go unchecked during
   lfsr_fs_gc.

2. If we're traversing a btree that gets mutated mid-traversal, we're
   kicked entirely off the btree. This means we could miss large ranges
   of btree nodes/data blocks that may not have themselves been mutated.
   Not great.

Worst case, it doesn't hurt to check things again if the filesystem
changes. If this is too much of a bottleneck, you should probably be
running gc in incremental mode anyways, which always starts a new
traversal on ckmeta/ckdata.

Checking for dirty/mutated doesn't really add that much code:

           code          stack
  before: 36240           2680
  after:  36256 (+0.0%)   2680 (+0.0%)