t: Moved lookahead population into lfsr_mtree_gc/lfs_alloc
It was a bit weird to have this in lfsr_mtree_traverse, which doesn't
change any filesystem state otherwise.

At the very least we should call lfs_alloc_markinuse and
lfs_alloc_markfree in the same function, and rerouting
lfsr_mtree_traverse eot to handle this would have added code cost
anyways.

The main cost is stack:

           code          stack
  before: 36256           2680
  after:  36288 (+0.1%)   2704 (+0.9%)

Unfortunately this reveals one of the bigger issues with our optional
return parameters: if a function with optional return parameters needs
the structs to perform work, in this case lfsr_mtree_traverse needs
lfsr_bptr_t in case ckmeta/ckdata is requested, it requires an
additional stack allocation.

In theory, these stack allocations could be elided if the return structs
are provided, but you can't really express this in standard C.

Combine this with the fact that lfs_alloc is sensitive to stack changes,
and lives at the bottom at every hot-path, and the end result is more
stack usage.

Note that the additional stack cost, 24 bytes, is exactly equal to one
tag + one bptr, 4 bytes + 20 bytes.