Reverted to simpler ecksum calculation
So simply calculating the ecksum over the whole prog size, instead of
manually CRCing the leading byte that we need to separately read to
check if we need to perturb.

Yes, this risks two reads instead of the one we need, but it's simpler,
less error prone, and less code. Our caching layer should prevent double
reads like this, so we might as well rely on it.

Saves some code:

           code          stack
  before: 36396           2664
  after:  36368 (-0.1%)   2664 (+0.0%)