Rearranged flags a bit
Mainly to make space for more shared open/mount flags that are future
planned.

The nice thing about our flags is they don't live on-disk, so we can
always change them whenever we need to.

It gets a bit messy, but this is what the current layout looks like:

              8     8     8     8
            .----++----++----++----.
            .----..-..-..----------.
  o_flags:  |type||f||t||    o     |
            |----||-|:-:'--.-.-----'
            |----||-|:-:---:-:-----.
  d_flags:  |type||f|: :   : :     |
            |----||-|:-:---:-:-----'
            |----||-|:-'--..-..----.
  t_flags:  |type||f|| t  ||f||tstt|
            '----''-'|----|'-''----'
            .--------|----|:-:-----.
  gc_flags: |        | t  |: :     |
            '--------|----|:-:-----'
            .----..-.|----|:-:.----.
  f_flags:  | f  ||m|| t  |: :| f  |
            '----'|-||----|:-:'----'
            .----.|-||----||-|.----.
  m_flags:  | i  ||m|| t  ||o|| m  |
            |----||-|'----'|-||----|
            |----||-|------|-||----|
  i_flags:  | i  ||m|      |o|| m  |
            '----''-'------'-''----'

The main downside of this layout is that some of the traversal flags,
LFS_T_DIRTY/MUTATED, now risk ambiguity with open/mount flags. But I
don't think this is really avoidable as traversals are already using
almost the entire 32-bit encoding space...

No code changes:

           code          stack
  before: 36480           2680
  after:  36480 (+0.0%)   2680 (+0.0%)