Added btree/branch fetch functions
These weren't really necessary when btree/branch fetch was nothing more
than tag decoding, but now that we have ckfetches it makes sense to
deduplicate things for a bit of code savings.

One reason I was punting on this was I wasn't really sure if btree/
branch fetch should take decoded fields or the raw lfsr_data_t. We don't
get any code savings with the former, but it's the only API that's
consistent with lfsr_mdir_fetch/lfsr_rbyd_fetch/etc. I was going to go
with lfsr_btree_fetch/fetch_, but fortunately shoving the latter into
the lfsr_data_* namespace solved this dilemma:

- lfsr_branch_fetch     - fetches from decoded fields
- lfsr_data_fetchbranch - fetches from raw data + weight
- lfsr_btree_fetch      - fetches from decoded fields
- lfsr_data_fetchbtree  - fetches from raw data

Unfortunately, lfsr_btree_parent creates a bit of a wrinkle. We don't
want to redundantly fetch the child we're looking for, so we need to
decode and fetch in separate steps. This prevents inlining between these
small functions that could otherwise take place.

And, while they do save a bit of code, the position of these fetch
functions in the stack hot-path end up increasing our total stack
usage when ckfetches are enabled:

                     code          stack
  default before:   36428           2680
  default after:    36456 (+0.1%)   2680 (+0.0%)

  ckfetches before: 36848           2680
  ckfetches after:  36686 (-0.4%)   2712 (+1.2%)

But maybe this is just indicative of us not accounting for
shrinkwrapping?