Moved info flags to overlap with traversal flags
We just have too many flags! Mount flags specifically are already close
to filling up with the currently planned features.

Fortunately the info flags, used internally to track filesystem state,
are never needed at the same time as the traversal flags which specify
one-time traversals during lfsr_mount. So we can move these to overlap
and free up quite a bit more space:

              8     8     8     8
            .----++----++----++----.
            .----..-..-..----------.
  o_flags:  |type||f||t||    o     |
            |----||-|:-:'--.-.-----'
            |----||-|:-:---:-:-----.
  d_flags:  |type||f|: :   : :     |
            |----||-|:-:---:-:-----'
            |----||-|:-'--..-..----.
  t_flags:  |type||f|| t  ||f||tstt|
            '----''-'|----|'-''----'
            .--------|----|:-:-----.
  gc_flags: |        | t  |: :     |
            '--------|----|:-:-----'
            .-------.|----|.-------.
  f_flags:  |   m   || t  ||   f   |
            |-------||----|'-------'
            |-------||----|:-:.----.
  m_flags:  |   m   || t  ||o|| m  |
            |-------|'----'|-||----|
            |-------|.----.|-||----|
  i_flags:  |   m   || i  ||o|| m  |
            '-------''----''-''----'

The only downside is a bit more masking and not having this info
available when debugging.

The overlap is also convenient for lfsr_fs_gc and lets us remove some
shifts, which humorously perfectly canceled out the added cost of the
masks:

           code          stack
  before: 36416           2616
  after:  36416 (+0.0%)   2616 (+0.0%)